requirements.txt
fastapi
uvicorn
ultralytics
scikit-learn
torch
torchvision
opencv-python-headless
python-multipart
numpy
pillow
